quiz_microservice/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
PyPDF2==3.0.1
chromadb==0.5.23
requests==2.32.3
python-decouple==3.8
